Save transactions to the database

+++ nattbakka-server/Models/Transaction.cs
@@ -5,7 +5,7 @@
         public int id { get; set; }
         public string tx { get; set; }
         public string address { get; set;}
-        public int sol { get; set;}
+        public double sol { get; set;}
         public bool sol_changed { get; set; }
         public int dex_id { get; set;}
         public int group_id { get; set;}
